+++ include/vrv/options.h
@@ -631,7 +631,6 @@ public:
     OptionBool m_condenseNotLastSystem;
     OptionBool m_condenseTempoPages;
     OptionBool m_evenNoteSpacing;
-    OptionString m_expand;
     OptionIntMap m_footer;
     OptionIntMap m_header;
     OptionBool m_humType;
@@ -733,6 +732,7 @@ public:
     OptionBool m_octaveAlternativeSymbols;
     OptionDbl m_octaveLineThickness;
     OptionBool m_octaveNoSpanningParentheses;
+    OptionDbl m_ossiaStaffSize;
     OptionDbl m_pedalLineThickness;
     OptionDbl m_repeatBarLineDotSeparation;
     OptionDbl m_repeatEndingLineThickness;
@@ -748,6 +748,7 @@ public:
     OptionBool m_spacingDurDetection;
     OptionDbl m_spacingLinear;
     OptionDbl m_spacingNonLinear;
+    OptionDbl m_spacingOssia;
     OptionInt m_spacingStaff;
     OptionInt m_spacingSystem;
     OptionDbl m_staffLineWidth;
@@ -771,9 +772,13 @@ public:
 
     OptionArray m_appXPathQuery;
     OptionArray m_choiceXPathQuery;
+    OptionString m_expand;
+    OptionBool m_expandFirst;
+    OptionBool m_expandGenerate;
     OptionBool m_loadSelectedMdivOnly;
     OptionBool m_mdivAll;
     OptionString m_mdivXPathQuery;
+    OptionBool m_ossiaHidden;
     OptionArray m_substXPathQuery;
     OptionString m_transpose;
     OptionJson m_transposeMdiv;