+++ src/doc.cpp
@@ -970,6 +970,11 @@ void Doc::ScoreDefSetCurrentDoc(bool force)
     ScoreDefSetCurrentFunctor scoreDefSetCurrent(this);
     this->Process(scoreDefSetCurrent);
 
+    if (scoreDefSetCurrent.HasOssia()) {
+        ScoreDefSetOssiaFunctor scoreDefSetOssia(this);
+        this->Process(scoreDefSetOssia);
+    }
+
     this->ScoreDefSetGrpSymDoc();
 
     m_currentScoreDefDone = true;
@@ -1601,14 +1606,38 @@ void Doc::TransposeDoc()
 
 void Doc::ExpandExpansions()
 {
-    // Upon MEI import: use expansion ID, given by command line argument
+    // Upon MEI import: use expansion ID or boolean for first expansion, given by command line argument
     std::string expansionId = this->GetOptions()->m_expand.GetValue();
-    if (expansionId.empty()) return;
+    bool expandFirst = this->GetOptions()->m_expandFirst.GetValue();
 
-    Expansion *startExpansion = dynamic_cast<Expansion *>(this->FindDescendantByID(expansionId));
-    if (startExpansion == NULL) {
-        LogWarning("Expansion ID '%s' not found. Nothing expanded.", expansionId.c_str());
-        return;
+    if (this->GetOptions()->m_expandGenerate.GetValue()) {
+        ExpansionMap expansionMap;
+        ListOfObjects scores = this->FindAllDescendantsByType(SCORE);
+        for (Object *object : scores) {
+            Score *score = vrv_cast<Score *>(object);
+            assert(score);
+            expansionMap.GenerateExpansionFor(score);
+        }
+        // Enable expansion of first one when generated
+        expandFirst = true;
+    }
+
+    if (expansionId.empty() && !expandFirst) return;
+
+    Expansion *startExpansion = NULL;
+    if (!expansionId.empty()) {
+        startExpansion = dynamic_cast<Expansion *>(this->FindDescendantByID(expansionId));
+        if (startExpansion == NULL) {
+            LogWarning("Expansion ID '%s' not found. Nothing expanded.", expansionId.c_str());
+            return;
+        }
+    }
+    else {
+        startExpansion = dynamic_cast<Expansion *>(this->FindDescendantByType(EXPANSION));
+        if (startExpansion == NULL) {
+            LogWarning("No expansion found. Nothing expanded.");
+            return;
+        }
     }
 
     xsdAnyURI_List existingList; // list of xml:id strings of elements already in the document