+++ src/view_page.cpp
@@ -44,6 +44,7 @@
 #include "mrest.h"
 #include "note.h"
 #include "options.h"
+#include "ossia.h"
 #include "page.h"
 #include "pageelement.h"
 #include "pagemilestone.h"
@@ -728,8 +729,7 @@ void View::DrawBarLines(DeviceContext *dc, Measure *measure, StaffGrp *staffGrp,
         // Get the corresponding staff
         AttNIntegerComparison comparison(STAFF, staffDef->GetN());
         Staff *staff = vrv_cast<Staff *>(measure->FindDescendantByComparison(&comparison, 1));
-        if (!staff) {
-            LogDebug("Could not get staff (%d) while drawing staffGrp - DrawBarLines", staffDef->GetN());
+        if (!staff || (staff->HasVisible() && (staff->GetVisible() == BOOLEAN_false))) {
             yBottomPrevious = VRV_UNSET;
             continue;
         }
@@ -1120,7 +1120,7 @@ void View::DrawMNum(DeviceContext *dc, MNum *mnum, Measure *measure, System *sys
     assert(measure);
     assert(mnum);
 
-    Staff *staff = system->GetTopVisibleStaff();
+    Staff *staff = system->GetTopVisibleStaff(true);
     if (staff) {
         // Only one FloatingPositioner on the top (visible) staff
         if (!system->SetCurrentFloatingPositioner(staff->GetN(), mnum, staff, staff)) {
@@ -1177,6 +1177,86 @@ void View::DrawMNum(DeviceContext *dc, MNum *mnum, Measure *measure, System *sys
 }
 
 //----------------------------------------------------------------------------
+// View - Ossia
+//----------------------------------------------------------------------------
+
+void View::DrawOssia(DeviceContext *dc, Ossia *ossia, Measure *measure, System *system)
+{
+    assert(dc);
+    assert(ossia);
+    assert(measure);
+    assert(system);
+
+    dc->StartGraphic(ossia, "", ossia->GetID());
+
+    const Staff *topOStaff = ossia->GetDrawingTopOStaff();
+    const Staff *bottomOStaff = ossia->GetDrawingBottopOStaff();
+
+    // Draw scoreDef line and brace
+    if (ossia->IsFirst() && ossia->DrawScoreDef() && ossia->HasMultipleOStaves()) {
+        if (topOStaff && bottomOStaff) {
+            const int staffSize = bottomOStaff->m_drawingStaffSize;
+            const int x = topOStaff->GetDrawingX() + topOStaff->GetOssiaDrawingShift(measure, m_doc);
+            const int y1 = topOStaff->GetDrawingY();
+            const int doubleUnit = m_doc->GetDrawingDoubleUnit(staffSize);
+            const int y2 = bottomOStaff->GetDrawingY() - doubleUnit * (bottomOStaff->m_drawingLines - 1);
+            // Bar lines always 100
+            const int barLineWidth = m_doc->GetDrawingBarLineWidth(100);
+            this->DrawVerticalLine(dc, y1, y2, x + barLineWidth / 2, barLineWidth);
+            this->DrawBrace(dc, x, y1, y2, staffSize);
+        }
+    }
+
+    for (Object *child : ossia->GetChildren()) {
+        if (child->Is(STAFF)) {
+            this->DrawStaff(dc, vrv_cast<Staff *>(child), measure, system);
+        }
+        else {
+            assert(false);
+        }
+    }
+
+    // No bar lines to draw if we have no ossia staves (e.g., all hidden)
+    if (!topOStaff || !bottomOStaff) {
+        dc->EndGraphic(ossia, this);
+        return;
+    }
+
+    bool showBarLines = ossia->HasShowBarLines() ? (ossia->GetShowBarLines() == BOOLEAN_true) : false;
+    bool showForceLeft = (showBarLines && ossia->IsFirst() && (measure->GetDrawingLeftBarLine() == BARRENDITION_NONE));
+
+    // Draw bar lines
+    if (measure->GetDrawingLeftBarLine() != BARRENDITION_NONE) {
+        int yBottomPrevious = VRV_UNSET;
+        BarLine *barLine = measure->GetLeftBarLine();
+        dc->StartGraphic(barLine, "", barLine->GetID());
+        this->DrawBarLines(dc, measure, ossia->GetDrawingStaffGrp(), barLine, measure->IsLastInSystem(),
+            system->IsLastOfMdiv(), yBottomPrevious);
+        dc->EndGraphic(barLine, this);
+    }
+    if ((showBarLines || !ossia->IsLast()) && (measure->GetDrawingRightBarLine() != BARRENDITION_NONE)) {
+        int yBottomPrevious = VRV_UNSET;
+        BarLine *barLine = measure->GetRightBarLine();
+        dc->StartGraphic(barLine, "", barLine->GetID());
+        this->DrawBarLines(dc, measure, ossia->GetDrawingStaffGrp(), barLine, measure->IsLastInSystem(),
+            system->IsLastOfMdiv(), yBottomPrevious);
+        dc->EndGraphic(barLine, this);
+    }
+
+    if (showForceLeft) {
+        int yBottomPrevious = VRV_UNSET;
+        // Use the ossia drawing barline it that case
+        BarLine *barLine = ossia->GetDrawingLeftBarLine();
+        dc->StartGraphic(barLine, "", barLine->GetID());
+        this->DrawBarLines(dc, measure, ossia->GetDrawingStaffGrp(), barLine, measure->IsLastInSystem(),
+            system->IsLastOfMdiv(), yBottomPrevious);
+        dc->EndGraphic(barLine, this);
+    }
+
+    dc->EndGraphic(ossia, this);
+}
+
+//----------------------------------------------------------------------------
 // View - Staff
 //----------------------------------------------------------------------------
 
@@ -1186,6 +1266,8 @@ void View::DrawStaff(DeviceContext *dc, Staff *staff, Measure *measure, System *
     assert(staff);
     assert(measure);
     assert(system);
+
+    if (staff->IsHidden()) return;
 
     assert(system->GetDrawingScoreDef());
     StaffDef *staffDef = system->GetDrawingScoreDef()->GetStaffDef(staff->GetN());
@@ -1253,6 +1335,10 @@ void View::DrawStaffLines(DeviceContext *dc, Staff *staff, StaffDef *staffDef, M
 
     x1 = measure->GetDrawingX();
     x2 = x1 + measure->GetWidth();
+    if (staff->IsOssia()) {
+        int shift = staff->GetOssiaDrawingShift(measure, m_doc);
+        x1 += shift;
+    }
     y1 = staff->GetDrawingY();
     if (!staff->HasDrawingRotation()) {
         y2 = y1;
@@ -1561,7 +1647,7 @@ void View::DrawSystemDivider(DeviceContext *dc, System *system, Measure *firstMe
 
     if ((system->IsDrawingOptimized() || (m_options->m_systemDivider.GetValue() > SYSTEMDIVIDER_auto))) {
         int y = system->GetDrawingY();
-        Staff *staff = system->GetTopVisibleStaff();
+        Staff *staff = system->GetTopVisibleStaff(true);
         if (staff) {
             // Place it in the middle of current and previous systems - in very tight layout this can collision with
             // the staff above. To be improved
@@ -1665,7 +1751,10 @@ void View::DrawMeasureChildren(DeviceContext *dc, Object *parent, Measure *measu
     }
 
     for (Object *current : parent->GetChildren()) {
-        if (current->Is(STAFF)) {
+        if (current->Is(OSSIA)) {
+            this->DrawOssia(dc, vrv_cast<Ossia *>(current), measure, system);
+        }
+        else if (current->Is(STAFF)) {
             // cast to Staff check in DrawStaff
             this->DrawStaff(dc, vrv_cast<Staff *>(current), measure, system);
         }