+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # Changelog
 
 ## [unreleased]
+* Support for ossia staves
+* Support for expansion generation in MusicXML importer
+* Improved MusicXML importer (articulations, spacing, clef changes, tempo spacing, coda, segno)
+* Improved `note@head.fill` support
+* Improved SVG styling structure
+* Option `--spacing-ossia` for adjusting the spacing of ossia staves
+* Option `--ossia-hidden` for hiding rendering of ossia staves
+* Option `--expand-first` to expand first `expansion` in the score
+* Option `--expand-generate` to generate `expansion` from measure repeats
+
+## [5.7.0] – 2025-12-02
 * Support for `@ho` and `@vo` on relevant layer elements and control events
 * Support for `@startho/vo` and `@endho/vo` on `hairpin`, `lv`, `phrase`, `slur` and `tie`
 * Support for `@enclose` on `rest`, `mordent`, `turn`, and `trill` (@rettinghaus)